LF-4696 Move useGroupedSensors under SensorList
