+++ packages/webapp/src/containers/SensorList/index.tsx
@@ -16,7 +16,7 @@
 import { History } from 'history';
 import Loading from '../../components/Form/ContextForm/Loading';
 import PureEsciSensors from '../../components/Sensor/v2/EsciSensors';
-import useGroupedSensors from '../Sensor/useGroupedSensors';
+import useGroupedSensors from './useGroupedSensors';
 import { partnerIds } from '../LocationDetails/PointDetails/SensorDetail/v2/constants';
 
 interface SensorListProps {
